Safari-611.1.10/LayoutTests/editing/inserting/insert-paragraph-separator-at-break.html

<!DOCTYPE html>
<html>
<body>
<div contenteditable id="test"><b>place the cursor between these two lines<br><br>and press return</b><br>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This test ensures WebKit inserts only a break element when inserting a paragraph at a break element.\n'+
    'Only one BR should be inserted in this test.');

var elem = document.getElementById('test');
window.getSelection().setPosition(elem.firstChild, 2);
document.execCommand("InsertParagraph");
Markup.dump(elem, "After");

</script>
</body>
</html>
