third_party/blink/web_tests/platform/generic/editing/inserting/insert-paragraph-separator-at-break-expected.txt

This test ensures WebKit inserts only a break element when inserting a paragraph at a break element.
Only one BR should be inserted in this test.

After:
| <b>
|   "place the cursor between these two lines"
|   <br>
|   <br>
|   <#selection-caret>
|   <br>
|   "and press return"
| <br>
| "
"
